:construction_worker: Add a re-usable OAS workflow
This workflow is composed of the reusable actions and should over
sane defaults, while allowing usage in a matrix build strategy for
projects with multiple API specs.

The jobs are organized in such a way that early failures prevent
doing useless work - e.g. generating a postman collection or SDKs from
an API spec that is outdated has no purpose, as that has already been
done in previous commits/the default branch.